moved get_settled_transactions to razorpay since it is payment gateway specific
